fix: resolve CI failures for Hytte-5tmc
