+++ web/src/pages/WorkHoursPage.tsx
@@ -6,25 +6,9 @@ import { Skeleton } from '../components/ui/skeleton'
 import { ConfirmDialog } from '../components/ui/dialog'
 import { Select, type SelectOption } from '../components/ui/select'
 import { TimePicker } from '../components/ui/time-picker'
+import { calculateDayWithLivePunch, type WorkSession, type WorkDeduction, type WorkSettings } from './workHoursUtils'
 
 // ── Interfaces ──────────────────────────────────────────────────────────────
-
-interface WorkSession {
-  id: number
-  day_id: number
-  start_time: string
-  end_time: string
-  sort_order: number
-  is_internal: boolean
-}
-
-interface WorkDeduction {
-  id: number
-  day_id: number
-  name: string
-  minutes: number
-  preset_id?: number | null
-}
 
 interface WorkDay {
   id: number
@@ -315,60 +299,6 @@ function currentTimeHHMM(): string {
   return `${String(now.getHours()).padStart(2, '0')}:${String(now.getMinutes()).padStart(2, '0')}`
 }
 
-export interface WorkSettings {
-  standard_day_minutes: number
-  lunch_minutes: number
-  rounding_minutes: number
-}
-
-export interface LiveEstimate {
-  grossMinutes: number
-  lunchMinutes: number
-  deductionMinutes: number
-  netMinutes: number
-  reportedMinutes: number
-  standardMinutes: number
-}
-
-export function calculateDayWithLivePunch(
-  now: Date,
-  punchStart: string,
-  sessions: WorkSession[],
-  lunch: boolean,
-  deductions: WorkDeduction[],
-  settings: WorkSettings,
-): LiveEstimate | null {
-  const [sh, sm] = punchStart.split(':').map(Number)
-  const startMins = sh * 60 + sm
-  const nowMins = now.getHours() * 60 + now.getMinutes()
-  if (nowMins < startMins) return null
-
-  let gross = nowMins - startMins
-  for (const s of sessions) {
-    const [sSh, sSm] = s.start_time.split(':').map(Number)
-    const [sEh, sEm] = s.end_time.split(':').map(Number)
-    gross += sEh * 60 + sEm - (sSh * 60 + sSm)
-  }
-
-  const lunchMin = lunch ? settings.lunch_minutes : 0
-  let customMin = 0
-  for (const d of deductions) {
-    customMin += d.minutes
-  }
-
-  const net = Math.max(gross - lunchMin - customMin, 0)
-  const rounding = settings.rounding_minutes > 0 ? settings.rounding_minutes : 30
-  const reportedMin = Math.floor(net / rounding) * rounding
-
-  return {
-    grossMinutes: gross,
-    lunchMinutes: lunchMin,
-    deductionMinutes: customMin,
-    netMinutes: net,
-    reportedMinutes: reportedMin,
-    standardMinutes: settings.standard_day_minutes,
-  }
-}
 
 // ── Main page ──────────────────────────────────────────────────────────────
 
@@ -581,7 +511,6 @@ function DayView({
 
   useEffect(() => {
     if (punchStart === null) return
-    setNow(new Date())
     const id = setInterval(() => setNow(new Date()), 60_000)
     return () => clearInterval(id)
   }, [punchStart])
